mender-binary-delta: Fix broken do_version_check.
For unknown reasons, the "noexec" flag is not actually possible to
turn off, not even with "0" nor "". The only way to disable it is to
remove it, which kind of defeats the purpose of having it. Luckily
it's easy to work around, so just do the variable check in code
instead.

The reason this has passed in our CI is that it is not covered by the
sstate cache, so if it had a hit then this problem would not reveal
itself.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>