[tests/acceptance] Remove --no-pull and update image-tests
From now on, the integration pipeline will not be cloning separately
mender-image-tests and then hacking meta-mender to use such clone, but
just use whatever is submodule'd here.

This commit updates the submodule to latest, removes the --no-pull
functionality, and removes the ignore of the submodule to make
developers aware.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>