Add PACKAGECONFIG option for the inventory-network-scripts target
This adds a PACKAGECONFIG option: 'inventory-network-scripts' to the
mender-client_% recipes, which enables controlling whether or not to install the
inventory-network scripts in the client (for now, this is only the 'inventory-geo-script').

Also, a simple build test is added to test_build.py.

In order to support the backwards compatibility of building with the old
'install-inventory-scripts' target, the do_install step has been copied to the
2.3.1 and 2.4.1 recipes, while the include file has the new install targets.

Changelog: New Mender-client configuration option: 'inventory-network-scripts'.
This option, if enabled, installs the inventory-network-scripts in the client.
This is enabled as an option, because the inventory-geo script relies on a
third-party network service to figure out the geographic location of the device,
which may not be something that everyone wants installed on their devices. The
feature is enabled in the standard 'PACKAGECONFIG' for the Mender-client, and is
included unless overridded. To remove it, add
'PACKAGECONFIG_remove_pn-mender-client = "inventory-network-scripts"' to your
local.conf file.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>