Fix parsing order so `inherit` works with `MENDER_FEATURES`.
It's two combined fixes that solve the problem: First, we need to move
the inclusion of `mender-setup` so that it comes after the definition
of all variables that add to `MENDER_FEATURES`. Second, we need to
populate `OVERRIDES` in a variable assignment instead of in an
anonymous Python function. The latter is executed at the end, whereas
assignment is executed when it is needed (early).

Changelog: Fix a parsing issue where `inherit` could not be used with
variables that had been defined with overrides that depended on
`MENDER_FEATURES_ENABLE`. One example would be:
```
MYVAR_mender-grub = "grub-efi"
inherit ${MYVAR}
```

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit d3e4b6edb12a155fc6a14ce876115686e64b31ee)

Backported-by: Kristian Amlie <kristian.amlie@northern.tech>
Backported from master using OVERRIDES instead of DISTROOVERRIDES.