Add missing dependencies for Python 3 requirements file
These were installed in the test image but not tracked in meta-mender.

Note that PyYAML is not an strict dependency for the tests, but a
requirement for our release_tool. We add it here for the sake of
simplicity.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit baf87988825ffe27f78cd2b73bf37a2f6d4b5874)