Fix mender-client 2.4.1/2.3.1 do_install tasks
They should use `MENDER_FEATURES` to check whether `mender-image`
feature is enabled or not.

Without this fix, the resulting image does not contain the required
symbolic link in /data partition.

This mistake was introduced at ab0e92e, likely during a merge/rebase
before final merge of the PR. The tasks are a copy of the .inc one
before introducing the inventory-network-scripts make target.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>