MEN-4294: Re-enable all of the test_dbus_fetch_jwt_token test.
Also fix one issue in the test: The bootstrap needs to happen before
the dbus monitor is started, because the bootstrap also broadcasts the
token. This will trigger the check for the broadcast too early, and we
will proceed with other DBus calls before the client is fully up and
running. By starting the monitor after the bootstrap, we know that the
client is up when we receive the signal.

Changelog: None
Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 983807883c12866b528da1b089f85dcb2149e35a)