Add meta-oe as dependency to mender-demo
`jq` is a dependency for mender configure in the demos.
It doesn't exit in poky only in meta-oe.

"meta-oe" is called the "openembedded-layer".

Trying to build without meta-oe produces this error:

```
ERROR: Nothing RPROVIDES 'jq' (but
/home/user/demo_notes/beaglebone-yocto/build/../sources/meta-mender/meta-mender-core/recipes-mender/mender-configure/mender-configure_1.0.0.bb
RDEPENDS on or otherwise requires it)
```

This commits makes it more explicit on how to fix the issue:

```
ERROR: Layer 'mender-demo' depends on layer 'openembedded-layer', but this layer is not enabled in your configuration
```

Changelog: Title

Signed-off-by: Alan <alan.martinovic@gmail.com>
(cherry picked from commit bac727fbb7600e8a1bd1ba085b8c1c6ab64eb114)