MEN-4550: Add acceptance test for DBus Update Control endpoint.
`mender-mock-server.py` has to be substantially extended for the calls
to work correctly. We have to consume the body of requests for
subsequent calls to work, and we need to be able to handle chunked
bodies. We also make it possible to schedule a deployment with it.

In addition we add a logger-update-module which we can use to track
exactly which states the Update Module is going through.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit f855270c68e50f7a9c5e0f95921ca677e830b3bc)
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>