Add the update control variables to the demo configuration
This adds the possibility to configure the udpate control configuration
variables in 'mender.conf'.

This is done through adding:

* MENDER_UPDATE_CONTROL_MAP_EXPIRATION_TIME_SECONDS
* MENDER_UPDATE_CONTROL_MAP_BOOT_EXPIRATION_TIME_SECONDS

variables, which, if set, are added to the rootfs local configuration file in
/etc/mender/mender.conf.

It also adds a default value for these variables in the demo layer, in which the
values are set to 120 and 60, respectively.

Not documenting it in purpose, as we favor the method of adding a custom
mender.conf file to the build instead of setting up even more MENDER_
variables.

Changelog: Increase Update Control Map timeout values in demo to avoid
Update Control malfunctioning on slow boards.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit bee2bd25acb17091ea67d59b6196f588a1c2ff5b)
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 80f35813984c77f66ef460ce599d37920bd47aaf)
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>