Add lmdb recipe
This is required for use with the mender-monitor client, which has a dependency
on the mdb_{dump,load} commands.

The recipe is a simple copy-paste from the lmdb recipe in meta-openembedded
master. But since it is not present in any of the releases we rely on, this will
have to be mirrored in our repository for compatibility with the old
releases (Dunfell, and Gatesgarth amongst them).

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 8d1af303c02640b61b5d9421e67ce15ffbca757c)