Skip old U-Boot workaround in newer ones, where it doesn't work.
In U-Boot 2020.10, `include/config_defaults.h` was removed. This was
the place we used to put definitions we cannot find a Kconfig option
for, so this doesn't work anymore for newer versions of U-Boot.
However, since only older versions require the one option which still
goes into that file, just omit that whole step if the file doesn't
exist.

Changelog: Auto patching: Fix error `Unable to add definition
MTDIDS_DEFAULT`, which happens on U-Boot v2020.10 and later.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>