Revert U-Boot patch which breaks vexpress-qemu-flash.
This is related to the function `flash_detect_cfi`. Before the
`chip_lsb` variable was added, this function configured the
`vexpress-qemu-flash` board correctly. But with the addition of that
variable and extra check, it now returns too early, with `chip_lsb`
set to 1, which does not work in the `vexpress-qemu-flash`
configuration. The exact details of why this fails are lost on me,
hence this full revert.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>