mender-monitor_git: define unique PV based on SRCREV
This will force yocto to rebuild from source every time that the
filename of the tarball changes.

Now the PV for the git recipe will be master-git${SRCREV}

This change aligns this recipe with other Mender components that are
build from source, though in this case SRCREV is extracted from the
filename and not from an actual Git repository.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>