u-boot: Fix auto-patch saving when recipe version contains `+`.
When the version contains a '+', it will be interpreted as a wildcard
character, since we are using extended regular expressions. So switch
to basic expressions instead, which will still work if we tweak the
first expression a little bit.

Changelog: u-boot: Fix save_mender_auto_configured_patch when recipe
version contains a `+`.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>