MEN-5137: Enable demo functionality for the monitor-client
This commit adds an append file for mender-monitor to the demo layer, which is
dynamically added when the meta-mender-commercial layer is present.

This enables the demo subsystems and monitors to be installed, enabling some
basic monitoring functionality out of the box.

Changelog: Title
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 02757eaaae8458e203225fc56e5b29a738d2efc9)