mender: introduce partlabel feature to find partition by their label
fstab and mender-client-resize-data-part use hardcoded path to the partition.

This can be an issue when you want to use the same image but on machine where the
disk enumeration is slightly different.

Example: one machine have a NVMe disk and the other a mSATA.
Or one machine have two hard drivers and you have installed mender on
the second one.

To avoid this we look for the correct partition based on the label device.

Note that the the grub.cfg and /data/mender/mender.conf should be correctly
configured by the user and should not rely on the image disk generated.

Changelog: Add feature to detect boot and data partition based on their label.

Signed-off-by: Clément Péron <peron.clem@gmail.com>