Remove the link to u-boot for AARCH64
These are the following changes:
* Remove `_MENDER_PART_IMAGE_DEPENDS_append_mender-grub_aarch64`:
It is been hardcoded to u-boot, as we are not sure what bootloader is
used as UEFI provider, as it could easily be EDK2.
* Remove inclusion of `version_logic.inc`:
The file assumes grub is booted from u-boot, but if using EDK2 then
it forces a version conflict.
* Remove `kernel-devicetree` from
`MACHINE_ESSENTIAL_EXTRA_RDEPENDS_append_mender-image_aarch64`:
It assumes to use u-boot, but when using EDK2 it may use ACPI tables.

Changelog: Title

Signed-off-by: Drew Reed <drew.reed@arm.com>
Signed-off-by: Vikas Katariya <vikas.katariya@arm.com>