mender-client_git: disable to work around upstream
Changelog: None

Due to changes in Yocto upstream, the python functions in
mender-client_git do not parse anymore. In order to retain a
minimal working state, that recipe and include are masked out
until a fix is applied. This needs to be removed then.

TODO: fix syntax for recipes-mender/mender-client/mender-client_git.inc

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>