Reinstate git recipe and fix now failing non-idiomatic var style.
According to Yocto experts, the variable expansion by '${}' reference
was never supposed to work inside Python expressions, and only did so
by accident. Switch to 'd.getVar()' style, since this now fails on
kirkstone and later.

Apparently it does work for the majority of cases though, and we have
*a lot* of these in Mender. I'm not going to fix all of them now
unless we encounter problems.

Changelog: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>