fix: Force building in separate build directory, even with `externalsrc`.
The problem with building in `${S}` is that files remain there after
a build, due to bad interoperability between the `externalsrc` and
`go` classes, the `clean` target does not clean this up. Therefore,
force the creation of a separate build directory always, even if it's
based on an `externalsrc` directory.

This was always needed for coverage analysis, so just steal the code
from there.

Changelog: Title
Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>