fix(mender-mock-server): add missing update_control_map endpoint
The tests we're previously passing only in the cases the control map on the
device did not need to be refreshed. In this case the client would not refresh
the map, and the missing endpoint was never detected.

However in any case the client did, the test would fail.

The fix is simple, add the `v2/deployments/{id}/update_control_map` endpoint to
the mock-server.

Changelog: None
Ticket: QA-384

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit ee9acf391324eb9ed1e615a7583087b9e0e4f726)
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>