fix: Set the image size to 1024 MiB for vexpress-qemu-arm
This is part of the migration to Ubuntu 22.04, which has a newer version of
qemu, in which the sd-card size has to be a multiple of two larger than the
running image (which as of writing is around 600MiB's).

The fix is to resize the image before running qemu, using the qemu-img command.

Changelog: None
Ticket: QA-390

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>
(cherry picked from commit 96bd4189e32ee9c6899c2acd946dc96f0822d37e)