feat: mender-bootstrap Artifact support
This adds the ability to generate Bootstrap Artifacts in the Mender-enabled
Yocto images.

This is done through running the bootstrap generation in all the `dataimg`
generation functions, and a TYPEDEP dependency, causing the rootfs image to be
generated first, so that the bootstrap Artifact can contain the checksum of the
rootfs image.

The functionality is included for all newer versions of the Mender client, but
removed and replaced with the old `mender-artifact-info` recipe for the 3.3.0
client.

For older branches this will have to be ported to all older recipes.

Note that this explicitly sorts the `test_bootstrap_artifact_install` to run
first in all instances, due to the way our test infra is set up.

Changelog: Add support for generating bootstrap Artifacts as a part of the move
to populating the client database on boot, as opposed to on the first update
install.

Ticket: MEN-2585

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>