fix: Update pseudo db when excluding content from rootfs
This commit is inspired the commit below from poky:
https://git.yoctoproject.org/poky/commit/?id=f85a4a1462bc2105f7c62f2b09c1d092c7677ada

Although the code is different, it is applying the exact same type of
fix, and therefore I'm also reusing (most of) their commit message
below:

---

To exclude content from the rootfs, the prefunc makes a copy (using
hardlinks if possible) of the rootfs directory and associated pseudo
db, then removes files & directories as needed. However if these files
and directories are removed using the python functions os.remove and
shutil.rmtree, the copied pseudo db will not be updated correctly. For
files copied from the original rootfs, if hardlinks were used
successfully when copying the rootfs this should mean that the
relevant inodes can't be reused and so the risk of pseudo aborts
should be avoided. However, this logic doesn't apply for directories
(as they can't be hardlinked) and so there remains some risk of inodes
being reused and the pseudo db becoming corrupted.

To fix this, use the 'rm' command under pseudo when removing files &
directories from the copied rootfs to ensure that the copied pseudo db
is updated.

Changelog: Fix occasional pseudo abort issue when building certain
images, especially in parallel. The symptom was this message:
```
abort()ing pseudo client by server request.
See https://wiki.yoctoproject.org/wiki/Pseudo_Abort for more details on this
```

Ticket: MEN-5857

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>