test(update_control): fix the issue with /run logs filling up mem
After a bit of digging, I discovered that when KVM is once again re-enabled in
our acceptance test jobs, the client would run out of memory, as a result of the
'systemd-run' job filling up the `/run`, and hence the running memory of the
QEMU instance.

The fix is simple, simply don't allow the job to log at a level below error.

Ticket: QA-414

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>