fix: missing dependency to mender-artifact-info for all mender-client versions
With the latest changes, dependency to mender-artifact-info was removed
for all mender-client versions besides 3.3.0. This causes a broken
OTA-feature as mender-client is not able to install any further artifact
when file /etc/mender/mender_artifact_info is missing.

This change reverts changes from commit
222532bf36d3ac195d62e11e5cfd06df0ae5a1c0. As i don't understand the
whole purpose of this commit, the change needs be checked if any other
functionality is broken now.

Signed-off-by: Ruben Schwarz <r.schwarz@sotec.eu>