test: Work around BusyBox mktemp which is less capable.
It requires that the XXXs be in the last component. Actually the Linux
man page also says this, so I guess it just works by accident in GNU.
We need the suffix, so make a temporary directory instead and put the
file inside that.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 51b56c4cac5986d24299087eb1495af80c1a8394)