test(test-connect): Fix the mender-connect spurious failure tests
This fixes the mender-connect spurious failure tests.

The issue turned out to be that the time is passed in between the host and the
QEMU instance, and is being used as a single source of truth.

This created a window of opportunity in which the test would get the system time
on the host, then pass it into the QEMU instance to use as a timestamp in
`since=<time>` when filtering the logs with `journald`.

Hence, mender-connect would actually behave, but log the correct output in the
window of time which arose in between the host time, and QEMU's time, and it was
simple filtered out.

The fix is dead simple: Use the QEMU instance time as the single source of truth

Ticket: QA-483

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>