feat: Add the mender-client-cpp build support
Ticket: MEN-5796

Changelog: Add the `mender` recipe in
`meta-mender-core/recipes/mender/mender-client/` while still keeping support for
the old golang client builds. The new `mender` recipe creates two packages.
`mender-auth` and `mender-update` to provide the C++ implementations of the
given functionality. The given client implementation can be chosen based on the
PREFERRED_PROVIDER_virtual/mender-client, by setting it to 'mender' for the C++
client, and 'mender-client' for the golang client. The default is set to the C++
client.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>