fix: Patch `mender-artifact` git recipe for golang 1.14
The incompatibilities are in the vendored dependencies of
`mender-artifact`. Two patches revert specific upstream changes that
introduced a dependency on golang newer than 1.14 and the two others are
direct patches on the source code.

Changelog: Patch `mender-artifact` _next_ recipe to be compatible with
`dunfell` which builds with golang 1.14.

Ticket: MEN-6142

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>