fix: remove splash IMAGE_FEATURE from demo layer
The IMAGE_FEATURE "splash" is unconditionally added to the build
if the meta-mender-demo layer is included. This is at least a bad
practice, and should be considered a bug for many types of devices
that do not have displays. Generally, a layer.conf should not
modify the build just by being included.

Changelog: Title
Ticket: None

Signed-off-by: TheYoctoJester <jester@theyoctojester.info>