fix: mv image files instead of install, to avoid corrupting the image.
coreutils 9.0 introduces a new behavior in the "install" tool.
When used to install a sparse file, it now looks for holes and
may insert new holes. Using "mv" instead of install ensures that
images do not lose significant information.
This fix is beneficial to the use of bmaptool to write images to
a device, especially if the image contains a FAT filesystem.

Changelog: Title
Ticket: None

Signed-off-by: Pascal Hache <hacpa@touchtunes.com>