refac(mender-client): Split C++ and Golang recipes completely.
Even though there is some merit to sharing things, it's getting to a
point where it's quite messy and confusing. Even though it's more code
it's better to make a clean split, and apply specific fixes to each
without worrying about compatibility with the other. This can also
make it easier to remove Golang support later.

The downside is that some fixes may need to be applied twice.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>