test: Fix kernel modules not being included correctly.
This seems to be because even though mender-client provides
mender-auth, corresponding `RDEPENDS` entries are not respected unless
mender-auth is a real package. So just revert to the old `${PN}` entry
for the Golang client.

Since the files are now almost completely duplicated, just make two
complete copies and stop sharing. At some point when 3.x goes EOL we
can remove the oldest one.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>