fix: Don't use `MENDER_FEATURES_ENABLE` variable directly.
We should use `MENDER_FEATURES` because this is the final result of
calculating the features, otherwise we can miss feature toggles that
have special handling.

This fixes an issue where `mender-update-install` was not added
correctly, leading to boot issues on RPi.

No changelog for this, because it was never in a tagged release.

On cherry-picking to `dunfell`, adapted the fix for `DISTRO_FEATURES`.
The reasoning applies the same way.

Changelog: None
Ticket: MEN-7039

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>
(cherry picked from commit 765ade2cf7e6747e766e10c586817eabc60df306)
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>