fix: Update IMAGE_FEATURES to new findings and expose more widely.
New testing has revealed that it is the `has_journal` flag which is
ultimately the problem, not the 64bit flag. I think the latter may
have enabled the problematic behavior with `has_journal` however,
which is on by default.

Changelog: Remove `has_journal` flag from mkfs.ext4 when
`read-only-rootfs` flag is enabled, which causes issues in certain
configurations. Specifically it solves these problems:
* Checksum is not stable which can cause issues with binary delta
  updates. This is an in-memory effect only, the physical checksum on
  disk is unaffected.
* Mender-snapshot can generate corrupted rootfs dumps, for the same
  reason as the previous point.

Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>