refac(mender-systemd): Do not automatically enable systemd anymore.
It does not look like it's possible to do it safely anymore. The old
approach with `DISTRO_FEATURES_BACKFILL` will not work on scarthgap
and later, because there is a separate variable `INIT_MANAGER`, which,
if set to `sysvinit` (the fault), it starts competing with us, using
`DISTRO_FEATURES_BACKFILL_CONSIDERED`.

The next logical step would then be to use `INIT_MANAGER` instead, but
this does not work either because it eventually leads to loading
`init-manager-${INIT_MANAGER}.inc`. This particular step is sensitive
to load ordering, and therefore our assignment comes to late to have
any effect, leading to the `sysvinit` variant being loaded, even if we
have set it to `systemd`. The mentioned load step comes from
`bitbake.conf`, and the only files loaded before that are `local.conf`
and `layer.conf` files. We can not (reasonably) utilize any of these
files for what we want to do, because feature selection comes much
later in the load order.

So we pick the safest variant, and ask the user to set their
`INIT_MANAGER` variable correctly instead. We could have made it
adaptive instead, and simply disabled the `mender-systemd` feature if
that `INIT_MANAGER` is not used. But since we have poor support for
this, and it's likely to break daemon mode, default to an error
message instead. `mender-systemd` can still be disabled manually,
which will get rid of the message.

Changelog: Enabling systemd as init manager is no longer done
automatically by the meta-mender layer. In order to re-enable it, put
`INIT_MANAGER = "systemd"` in your `local.conf` file. Alternatively,
use `MENDER_FEATURES_DISABLE = "mender-systemd"`, but be aware that
this can break daemon mode.

Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>