fix: meta-mender-ci: resolve usrmerge conflict
meta-mender-ci contains the lsb-test recipe which serves as an
lsb_release dropin for Zeus and later releases. So far it was
installed to $(bindir) and $(base_bindir), which are identical
if the usrmerge DISTRO_FEATURE is active.

To resolve the conflict, install to $(bindir) only. No compatibility
considerations are taken as this is a Mender CI specific peculiarity.

Changelog: Title
Ticket: None

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>