test: Fix and future-proof `test_incorrect_Kconfig_setting`
The test fix is to use absolute path, so that the generated patch is
actually found by bitbake. See related commit 34cc1e4 which applies the
same fix to a different test.

Now, presumably the `bitbake` commands have been failing *for the wrong
reasons* since end of 2022 but the test was broken in the first place so
it went unnoticed.

Fixes:
* Set capture=True so that the output is, well, captured
* Assume `e.output` exists, otherwise please fail the test!
* Add an extra catch all for exceptions, so that we make sure we verify
  the correct error message.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>