test: Remove `awscli` from OS requirements
This package does not exist anymore in recent Ubuntu, where awscli has
been moved to an snap. In any case it is not used by the tests
themselves, I believe it is a leftover when we used to run the whole
pipeline with the same VM image (based on this OS requirements).

Add `file` which is a requirement for bitbake, and which was before
being indirectly installed with `awscli` package.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>