chore: use `do_compile` and `do_install` from go.bbclass
Our `do_compile()` and `do_install()` functions are doing almost the
exact same thing as the default ones from `go.bbclass`, so there is
really no reason to override them with our own.

By utilizing the already-existing functions, we get the go build flags,
which in turn fixes the `textrel` and `buildpaths` warnings in the
`mender-connect` and `mender-snapshot` recipes, because they already
contain `-trimpath` and `buildmode=pie`.

Ticket: MEN-7275

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>