feat: add mender-prepopulate-inactive-partition feature
Add mender-prepopulate-inactive-partition MENDER_FEATURE.
Disabled by default for more efficient compression, it can be
enabled to have the inactive partition populated with the same
contents of the active one.

This allows to check the bootloader integrations (A/B switching,
A/B update, A/B rollback) directly on the generated images
without additional modifications or processing.

Ticket: None
Changelog: Commit

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>