test: Swap order of "pn-" prefix to workaround SPDX/sstate issue
Setting/unsetting PREFERRED_VERSION and rebuilding the recipes
invalidates SPDX data for them.

The recipes that use externalsrc will correctly recreate the data the
next time a rootfs is being built, but for some inexplicable reason this
does not happen for recipes that using sources from upstream, an instead
the build will rely on the stale `do_create_runtime_spdx_setscene` from
before the tests.

At the time of writing this, this is the case for `mender-flash` so that
consecutive tests would fail with:
```
ERROR: No SPDX file found for package mender-flash, False sstate:mender-flash:core2...
```

But the really interesting bit is that this happens only if we build
first with `PREFERRED_VERSION:recipe` and then
`PREFERRED_VERSION:pn-recipe`, but not if done in the reverse order (??)
nor if build only for one of the two variants (??).

Note that we have already tried cleaning everything (too aggressive),
not using the cache (it didn't work) and manually recreating the SPDX
data (it didn't work):
* https://github.com/mendersoftware/meta-mender/pull/2181/commits/c66217fe4
* https://github.com/mendersoftware/meta-mender/pull/2190
* https://github.com/mendersoftware/meta-mender/pull/2199

Ticket: QA-784

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>