fix: package artifact-verify-key.pem in mender-update
The path ${sysconfdir}/mender/artifact-verify-key.pem was not
explicitly added to a FILES:${} variable, and therefore added
to the default ${PN}-named package. As that is derived from the
recipe file name, it ended up as the only file in the "mender"
package, yielding artifact signature verification in
mender-update broken.

This adds the verification key to the "mender-update" package.

Changelog: Title
Ticket: MEN-7204

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>