feat: add variable MENDER_UBOOT_MMC_ENV_LINUX_DEVICE_PATH_REDUND
Following this approach to storing the uBoot environment in the eMMC boot
partitions on boards that support it, a second variable needs to be added and
accounted for in the generation of fw_env.config since the eMMC boot partitions
appear as separate devices.
https://hub.mender.io/t/make-mender-use-emmc-boot-partitions-for-saving-uboot-environment/5457

Ticket: None
Changelog: Support setting the u-boot redundant environment in a
separate storage with the variable
`MENDER_UBOOT_MMC_ENV_LINUX_DEVICE_PATH_REDUND`. This feature is
backwards compatible.

Signed-off-by: Benjamin Byholm <benjamin.byholm@walkbase.com>