test: check for topology in dataimg
We need to add the topology for the build to succeed, but the
topology file is one the data parition, and not in the ext4 being
tested. Therefore we need to add core-image*.dataimg and check for the
topolgy there.

We were also adding passing it to the wrong argument to `build_image`,
so it wasn't properly added to the local.conf

Ticket: QA-1186

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>