test: fix mender-orchestrator-support test
Even though mender-orchestrator-support technically is open-source, it
requires a commercial component, so it's added under to
mender-commercial; therefore the test is moved to
`tests/acceptance/commercial`. Furthermore, it requires mender-orchestrator
(RDEPENDS), so we'll need to add it to IMAGE_INSTALL:append and add the
topology as a result. This will make the test require `SRC_URI:pn-mender-orchestrator`
to be set, but we don't require extra explicit tests, as that is already
covered by the mender-orchestrator recipe.

Ticket: QA-1186

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>