feat: add support for MENDER_DEVICE_TIER
Add an optional DeviceTier to mender-client-cpp.inc. This follows the
same structure as the TenantToken.

This will in this first iteration be used by mender-update to poll for
deployments and to install a manifest using the system_type defined in
mender-orchestrator's topology.

Ticket: MEN-8650
Changelog: Add support for specifying MENDER_DEVICE_TIER
in the mender conf.

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>