test: Split snapshot tests into two classes
To enable handling specific conditions for the mender-artifact ones.

By grouping also the fixtures (helps readability) a tiny detail modified
is that for the mender-artifact tests we require now client 2.5. This is
in any case so old that it is irrelevant for the tests nowadays.

Ticket: QA-1323

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 0cfde8139adce85637023468edf6bcd140ef190f)