test: Improve robustness of snapshot tests
There is a known issue with `mender-artifact` using the snapshot feature
in that the tool to detect the filesystem type may fail. We don't really
know the root cause, though.

The filesystem detection is then used to invoke yet another tool `fsck`
to check the filesystem. This whole process is actually optional,
because when `blkid` is not found `mender-artifact` just continues with
no triggering the check.

Anyway... just write a rerun filter for these specific tests that
sporadically fail in CI.

Note that the rerun filter only works on the non terminal version, when
using `screen` terminal multiplexer the command exit code is lost so the
rerun won't be applied (but then, this was already the case and this
test could not fail by design).

Ticket: QA-1323

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit bd6edca6b796a0ff0daf22f7120b231ae40b6d6c)