chore: add device-tier option to virtual device
`mender-client-qemu` is used as a base for the mender-orchestrator
virtual device used in tests, and it requires the device tier to be set
to `system` to function properly. This makes it easy to configure it.

Ticket: MEN-8844

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>
(cherry picked from commit eb3670c69c87b6a37b3b31740451f996c2e8e8b1)