chore:  Add a warning about `mender` v5 upgrade
The cherry-pick applied cleanly, but the changelog was rewritten to
consider the specifics of `kirkstone`.

Changelog: **Warning**: Forced upgrade to `mender` v5 when selecting
`mender` recipe for the client provider but keeping default versions.
There was an oversight at the time of releasing `mender` 5.0.2 in
`kirkstone` which, in effect, sets the default `mender` recipe for
builds using `mender` provider to the v5 series, forcing a major upgrade
on users that did select `mender` for the client provider and did not
pin `mender` to an specific version or series. This change adds a
warning. A revert was considered but deemed more risky, as forcing a
downgrade is a known path for breakage

Ticket: MEN-8977

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>
(cherry picked from commit 96c97dc6a3be5356d1c6f73d4f4f4ff12828cc50)
Signed-off-by: Lluis Campos <lluis.campos@northern.tech>