chore:  Add a warning about `mender` v5 upgrade
Changelog: **Warning**: Forced upgrade to `mender` v5 when using default
versions. There was an oversight at the time of releasing `mender` 5.0.2
in `scarthgap` which, in effect, sets the default `mender` recipe for
the build to the v5 series, forcing a major upgrade on users that did
not pin `mender` to an specific version or series. This change adds a
warning. A revert was considered but deemed more risky, as forcing a
downgrade is a known path for breakage

Ticket: MEN-8977

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>