fix: Add missing `curl` dependency to `mender-monitor` recipes
It is a required recipe. It went unnoticed for so long because users
(and our tests) typically use this add-on in combination with other
add-ons that have the dependency or any other software that depends on
it.

Added also `dbus` for correctness. This one is handled by `mender-auth`
but in any case `mender-connect` does direct D-Bus calls so it should
have it explicitly on its dependencies.

Ticket: MEN-8986

Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>