feat: Add `mender-client-version-inventory-script` recipe
This new recipe provides an inventory script with the "built-in"
inventory key `mender_client_version` for a supported Mender Client
installation.

It does this by conflicting with unsupported combination of versions
for any Mender Client component, so that we prevent users from
installing them.

The 6.0.0 recipe contained in this commit is a proof-of-concept of the
tentative first release together with the next Mender Client.

Changelog: Title
Ticket: MEN-9018

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>